refactor(esl-utils): separate scroll file
